+++ app/models/ckeditor/asset.rb
@@ -1,8 +1,11 @@
-# CKEditor base asset file
+# frozen_string_literal: true
+
 class Ckeditor::Asset < ActiveRecord::Base
   include Ckeditor::Orm::ActiveRecord::AssetBase
+  # include Ckeditor::Backend::ActiveStorage # TODO: AS-SWITCH
 
   delegate :url, :current_path, :content_type, to: :data
+  # attr_accessor :data # TODO: AS-SWITCH
 
-  validates_presence_of :data
+  has_one_attached :storage_data
 end